create controllers for user
