+++ backend/routes/advertRoutes.js
@@ -1,0 +1,1 @@
+// An authenticated endpoint which only returns content if the Authorization header with the user's token was correct
